requirements.txt
fastapi==0.135.1
openai==2.21.0
uvicorn==0.42.0
python-dotenv==1.1.0
